+++ src/api/rpc/pool/index.ts
@@ -66,7 +66,7 @@ export type TLoginRequest = {
 export type TLoginResponse = {
   farmer_record: FarmerRecord;
   recent_partials: Array<[uint64, uint64]>; // Array of tuple [timestamp, difficulty]
-} | { };
+} | Record<string, never>;
 export async function login(agent: RPCAgent, data: TLoginRequest){
   return agent.request<TLoginResponse>("GET", "login", data);
 }